requirements_gradio.txt
gradio>=4.0.0
python-dotenv>=1.0.0
telethon>=1.34.0
langchain>=0.1.0
langchain-community>=0.0.13
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openvino>=2024.2.0
openvino-tokenizers>=2024.2.0
openvino-genai>=2024.2.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml[html_clean]>=5.1.0 